Feat(13): 통합 로그인 (#15)
* Chore : 라이브러리 추가

* Feat : auth 설정

* Feat : 토큰 갱신, 삭제

* Feat : 카카오 로그인, 구글 로그인

* Style : 코드 포맷팅

* Refactor : 불필요한 코드 삭제

* Refactor : 매직넘버 제거

* Chore: docker compose db 파일 추가

* Feat : 통합 로그인 기능

통합 로그인 연동
통합 로그인

* Style : 코드 포맷팅

+++ src/main/java/com/quantweb/springserver/auth/support/kakao/dto/KakaoUserResponse.java
@@ -2,6 +2,7 @@ package com.quantweb.springserver.auth.support.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.HashMap;
 import lombok.Getter;
@@ -14,7 +15,7 @@ public class KakaoUserResponse {
 
   //회원 번호
   @JsonProperty("id")
-  public Long id;
+  public BigInteger id;
 
   //자동 연결 설정을 비활성화한 경우만 존재.
   //true : 연결 상태, false : 연결 대기 상태
@@ -93,7 +94,6 @@ public class KakaoUserResponse {
     public Boolean isAgeAgree;
 
     //연령대
-    //참고 https://developers.kakao.com/docs/latest/ko/kakaologin/rest-api#req-user-info
     @JsonProperty("age_range")
     public String ageRange;
